migrate routes, remove statistics module
