--- src/app/app-routing.module.ts
@@ -1,66 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { NotFoundComponent } from './core/not-found.component';
-import { LoadingGuard } from './shared/LoadingGuard';
-
-export const LOGIN_URL = '/auth/login';
-export const LOGGED_OUT_URL = '/auth/logged-out';
-
-const routes: Routes = [
-  {
-    path: '',
-    canActivate: [LoadingGuard],
-    data: {},
-    children: [
-      {
-        path: 'individuals',
-        loadChildren: () => import('./individual/individual.module').then(m => m.IndividualModule)
-      },
-      {
-        path: 'stations',
-        loadChildren: () => import('./station/station.module').then(m => m.StationModule)
-      },
-      {
-        path: 'map',
-        loadChildren: () => import('./map/map.module').then(m => m.MapModule)
-      },
-      {
-        path: 'statistics',
-        loadChildren: () => import('./statistics/statistics.module').then(m => m.StatisticsModule)
-      },
-      {
-        path: 'profile',
-        loadChildren: () => import('./profile/profile.module').then(m => m.ProfileModule)
-      },
-      {
-        path: '',
-        redirectTo: '/map',
-        pathMatch: 'full'
-      }
-    ]
-  },
-  {
-    path: 'auth',
-    children: [
-      {
-        path: '',
-        loadChildren: () => import('./login/login.module').then(m => m.LoginModule)
-      }
-    ]
-  },
-  {
-    path: 'login',
-    redirectTo: LOGIN_URL,
-    pathMatch: 'full'
-  },
-  {
-    path: '**',
-    component: NotFoundComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes, {})],
-  exports: [RouterModule]
-})
-export class AppRoutingModule {}
