+++ src/app/app.config.ts
@@ -3,18 +3,66 @@ import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http'
 import localeDe from '@angular/common/locales/de';
 import localeFr from '@angular/common/locales/fr';
 import localeIt from '@angular/common/locales/it';
-import { APP_INITIALIZER, ApplicationConfig, ErrorHandler, importProvidersFrom } from '@angular/core';
+import { APP_INITIALIZER, ApplicationConfig, ErrorHandler, enableProdMode, importProvidersFrom } from '@angular/core';
 import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
 import { getAuth, provideAuth } from '@angular/fire/auth';
 import { getFirestore, provideFirestore } from '@angular/fire/firestore';
-import { PreloadAllModules, provideRouter, Router, withDebugTracing, withPreloading } from '@angular/router';
+import { getStorage, provideStorage } from '@angular/fire/storage';
+import { ReactiveFormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
+import { provideAnimations } from '@angular/platform-browser/animations';
+import { Router, provideRouter } from '@angular/router';
 import { MissingTranslationHandler, TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import * as Sentry from '@sentry/angular-ivy';
-import { from, Observable } from 'rxjs';
+import { Integrations } from '@sentry/tracing';
+import { Observable, from } from 'rxjs';
+import { environment } from 'src/environments/environment';
 import { routes } from './app.routes';
 import { httpInterceptorProviders } from './http-interceptors';
+import { IndividualService } from './individual/individual.service';
+import { LoginModule } from './login/login.module';
+import { MapService } from './map/map.service';
+import { MasterdataService } from './masterdata/masterdata.service';
+import { UserService } from './profile/user.service';
+import { SensorsService } from './sensors/sensors.service';
 import { GlobalErrorHandler } from './shared/GlobalErrorHandler';
 import { SentryMissingTranslationHandler } from './shared/SentryMissingTranslationHandler';
+
+Sentry.init({
+  enabled: environment.sentryEnabled,
+  release: environment.version,
+  environment: environment.name,
+  dsn: 'https://b0f9e54dab264d1881553cbfbcc1641a@o510696.ingest.sentry.io/5606738',
+  autoSessionTracking: true,
+  integrations: [
+    new Integrations.BrowserTracing({
+      tracingOrigins: ['localhost', /^\//],
+      routingInstrumentation: Sentry.routingInstrumentation
+    })
+  ],
+  tracesSampleRate: environment.sentrySamplerate
+});
+
+Sentry.addGlobalEventProcessor(event => {
+  if (event.type === 'transaction') {
+    //remove specific ids to enable performance tracking in sentry
+    event.transaction = sanitizeTransactionName(event.transaction);
+  }
+  return event;
+});
+
+function sanitizeTransactionName(transaction: string) {
+  const regexp = new RegExp('/(individuals|stations|profile)/(?!new)([^/]*)');
+  if (regexp.test(transaction)) {
+    return transaction.replace(regexp, '/$1/:id');
+  } else return transaction;
+}
+
+if (environment.production) {
+  enableProdMode();
+
+  window.console.log = () => {};
+}
 
 export class CustomTranslateLoader implements TranslateLoader {
   getTranslation(lang: string): Observable<unknown> {
@@ -26,42 +74,21 @@ registerLocaleData(localeDe, 'de');
 registerLocaleData(localeFr, 'fr');
 registerLocaleData(localeIt, 'it');
 
-function getConfig() {
-  let config;
-  let request = new XMLHttpRequest();
-  try {
-    request.open('GET', '/__/firebase/init.json', false); // `false` makes the request synchronous
-    request.send(null);
-    if (request.status === 200) {
-      console.log(request.responseText);
-      config = request.responseText;
-    }
-
-    return JSON.parse(config);
-  } catch (e) {
-    console.error('environment:getConfig: unable to get api key : ', e);
-  }
-
-  return config;
-}
-
-// function loadConfig() {
-//   fetch('/__/firebase/init.json')
-//     .then(response =>
-//       response.json().then(config =>
-//         // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-//         platformBrowserDynamic([{ provide: FIREBASE_OPTIONS, useValue: config }]).bootstrapModule(AppModule)
-//       )
-//     )
-//     .catch(() => Sentry.captureMessage('Could not fetch firebase config.', 'fatal'));
-// }
-
-// void loadConfig();
-
-console.log('config', getConfig());
-
 export const appConfig: ApplicationConfig = {
   providers: [
+    provideRouter(routes),
+    importProvidersFrom(
+      BrowserModule,
+      ReactiveFormsModule,
+      TranslateModule.forRoot({
+        loader: {
+          provide: TranslateLoader,
+          useClass: CustomTranslateLoader
+        },
+        missingTranslationHandler: { provide: MissingTranslationHandler, useClass: SentryMissingTranslationHandler }
+      }),
+      LoginModule
+    ),
     {
       provide: ErrorHandler,
       useClass: GlobalErrorHandler
@@ -72,28 +99,22 @@ export const appConfig: ApplicationConfig = {
     },
     {
       provide: APP_INITIALIZER,
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
       useFactory: () => () => {},
       deps: [Sentry.TraceService],
       multi: true
     },
     httpInterceptorProviders,
-    // ScreenTrackingService,
-    // UserTrackingService,
+    MasterdataService,
+    UserService,
+    IndividualService,
+    SensorsService,
+    MapService,
+    provideFirebaseApp(() => initializeApp(environment.firebaseConfig)),
+    provideAuth(() => getAuth()),
+    provideStorage(() => getStorage()),
+    provideFirestore(() => getFirestore()),
     // { provide: DEBUG_MODE, useValue: true }
     provideHttpClient(withInterceptorsFromDi()),
-    importProvidersFrom(
-      TranslateModule.forRoot({
-        loader: {
-          provide: TranslateLoader,
-          useClass: CustomTranslateLoader
-        },
-        missingTranslationHandler: { provide: MissingTranslationHandler, useClass: SentryMissingTranslationHandler }
-      })
-    ),
-    provideFirebaseApp(() => initializeApp(getConfig())),
-    provideAuth(() => getAuth()),
-    provideFirestore(() => getFirestore()),
-    provideRouter(routes, withPreloading(PreloadAllModules), withDebugTracing())
+    provideAnimations()
   ]
 };
